Move attribute expansions to intercom-common
In theory this should allow writing tests for those expansions.
