+++ intercom-common/src/lib.rs
@@ -1,4 +1,5 @@
 #![feature(non_exhaustive)]
+#![recursion_limit="128"]
 
 #[macro_use] extern crate quote;
 #[macro_use] extern crate syn;
@@ -23,3 +24,4 @@ pub mod methodinfo;
 pub mod model;
 pub mod builtin_model;
 pub mod foreign_ty;
+pub mod attributes;
